+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 ### Added
 - log: Support de-facto standard `NO_COLOR` environment variable by disabling
   ANSI colors when this variable is set with non-empty value (#45).
+- settings: Support parameters deprecation (#48).
 
 ### Changed
 - log: Disable debug by default with `DaemonFormatter` similarly to